backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
httpx==0.26.0
pytest==8.0.0
email-validator==2.1.0.post1
python-multipart==0.0.9

# Phase 2: Multi-Agent & RAG dependencies
langgraph>=0.1.0
langchain>=0.1.4
langchain-google-genai>=0.0.6
chromadb>=0.5.0
sentence-transformers>=2.3.1
spacy>=3.7.2
rank-bm25==0.2.2
prometheus-client==0.20.0
pytest-asyncio>=0.24.0
aiohttp>=3.10.0
